requirements.txt
appdirs==1.4.3
Django==1.11
django-zxcvbn-password==2.0.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
stripe==1.53.0
zxcvbn-python==4.4.15

